+++ src/scandeval/model_setups/hf.py
@@ -1,5 +1,6 @@
 """Model setup for Hugging Face Hub models."""
 
+import importlib.util
 import logging
 import os
 import warnings
@@ -26,13 +27,14 @@ from ..exceptions import (
     FlashAttentionNotInstalled,
     HuggingFaceHubDown,
     InvalidBenchmark,
+    InvalidModel,
+    NeedsExtraInstalled,
     NoInternetConnection,
 )
 from ..languages import get_all_languages
 from ..protocols import GenerativeModel, Tokenizer
 from ..utils import (
     GENERATIVE_MODEL_TASKS,
-    HiddenPrints,
     block_terminal_output,
     create_model_cache_dir,
     get_class_by_name,
@@ -146,7 +148,7 @@ class HFModelSetup:
 
             # Check that the model exists. If it does not then raise an error
             if len(models) == 0:
-                raise InvalidBenchmark(
+                raise InvalidModel(
                     f"The model {model_id} does not exist on the Hugging Face Hub."
                 )
 
@@ -158,9 +160,9 @@ class HFModelSetup:
             elif "jax" in tags:
                 framework = Framework.JAX
             elif "spacy" in tags:
-                raise InvalidBenchmark("SpaCy models are not supported.")
+                raise InvalidModel("SpaCy models are not supported.")
             elif "tf" in tags or "tensorflow" in tags or "keras" in tags:
-                raise InvalidBenchmark("TensorFlow/Keras models are not supported.")
+                raise InvalidModel("TensorFlow/Keras models are not supported.")
 
             model_task: str | None = models[0].pipeline_tag
             if model_task is None:
@@ -223,6 +225,9 @@ class HFModelSetup:
                 and self.benchmark_config.device == torch.device("cuda")
             )
 
+        if load_in_4bit and importlib.util.find_spec("bitsandbytes") is None:
+            raise NeedsExtraInstalled(extra="generative")
+
         bnb_config = (
             BitsAndBytesConfig(
                 load_in_4bit=load_in_4bit,
@@ -256,6 +261,9 @@ class HFModelSetup:
             and os.getenv("USE_VLLM", True)
         )
 
+        if use_vllm and importlib.util.find_spec("vllm") is None:
+            raise NeedsExtraInstalled(extra="generative")
+
         if use_vllm:
             try:
                 model = VLLMModel(
@@ -270,7 +278,7 @@ class HFModelSetup:
                 # informative error message
                 oom_error_message = "No available memory for the cache blocks"
                 if oom_error_message in str(e):
-                    raise InvalidBenchmark("The model is too large to load on the GPU.")
+                    raise InvalidModel("The model is too large to load on the GPU.")
 
                 if self.benchmark_config.raise_errors:
                     raise e
@@ -280,9 +288,8 @@ class HFModelSetup:
                 use_vllm = False
                 logger.info(
                     "Failed to benchmark with vLLM - trying with the Hugging Face "
-                    "implementation instead."
+                    f"implementation instead. The error raised was {e!r}"
                 )
-                logger.debug(f"The error was: {e!r}")
 
         if not use_vllm:
             model_kwargs = dict(
@@ -330,7 +337,7 @@ class HFModelSetup:
                     if config.model_type == "deberta-v2":
                         config.pooler_hidden_size = config.hidden_size
 
-                    with warnings.catch_warnings(), HiddenPrints():
+                    with warnings.catch_warnings():
                         warnings.filterwarnings("ignore", category=UserWarning)
                         warnings.filterwarnings("ignore", category=FutureWarning)
                         try:
@@ -352,11 +359,11 @@ class HFModelSetup:
                                 model_kwargs["ignore_mismatched_sizes"] = True
                                 continue
                             else:
-                                raise InvalidBenchmark(str(e))
+                                raise InvalidModel(str(e))
                         except (TimeoutError, RequestError):
                             attempts_left -= 1
                             if attempts_left == 0:
-                                raise InvalidBenchmark(
+                                raise InvalidModel(
                                     "The model could not be loaded after 5 attempts."
                                 )
                             logger.info(
@@ -478,12 +485,12 @@ class HFModelSetup:
                 return config
             except KeyError as e:
                 key = e.args[0]
-                raise InvalidBenchmark(
+                raise InvalidModel(
                     f"The model config for the model {model_id!r} could not be "
                     f"loaded, as the key {key!r} was not found in the config."
                 )
             except OSError as e:
-                raise InvalidBenchmark(
+                raise InvalidModel(
                     f"Couldn't load model config for {model_id!r}. The error was "
                     f"{e!r}. Skipping"
                 )
@@ -528,7 +535,7 @@ class HFModelSetup:
                         trust_remote_code=self.benchmark_config.trust_remote_code,
                     )
                 except (JSONDecodeError, OSError, TypeError):
-                    raise InvalidBenchmark(
+                    raise InvalidModel(
                         f"Could not load tokenizer for model {model_id!r}."
                     )
                 except (TimeoutError, RequestError):
@@ -539,16 +546,14 @@ class HFModelSetup:
     @staticmethod
     def _handle_loading_exception(exception: Exception, model_id: str) -> None:
         if "checkpoint seems to be incorrect" in str(exception):
-            raise InvalidBenchmark(
-                f"The model {model_id!r} has an incorrect checkpoint."
-            )
+            raise InvalidModel(f"The model {model_id!r} has an incorrect checkpoint.")
         if "trust_remote_code" in str(exception):
-            raise InvalidBenchmark(
+            raise InvalidModel(
                 f"Loading the model {model_id!r} needs to trust remote code. "
                 "If you trust the suppliers of this model, then you can enable "
                 "this by setting the `--trust-remote-code` flag."
             )
-        raise InvalidBenchmark(
+        raise InvalidModel(
             f"The model {model_id} either does not exist on the Hugging Face "
             "Hub, or it has no frameworks registered, or it is a private "
             "model. If it *does* exist on the Hub and is a public model then "